Implemented TouchListeners for MapTools.

+++ app/src/main/java/com/cerealkillers/rootrunner/GameObjects/MapToolBehavior.java
@@ -5,5 +5,5 @@ package com.cerealkillers.rootrunner.GameObjects;
  */
 public interface MapToolBehavior
 {
-    public void use(int x, int y);
+    public void use(float x, float y);
 }
